Separated out buy functions into buyGear, buyArmoire, and buyPotion
